fix: compile errors after upstream refactor
