+++ core/edr-cache-core/build.gradle.kts
@@ -28,6 +28,7 @@ dependencies {
 
     implementation(project(":spi:edr-spi"))
 
+    testImplementation(libs.edc.lib.query)
     testImplementation(testFixtures(project(":spi:edr-spi")))
     testImplementation(libs.edc.core.connector)
