+++ edc-tests/runtime/iatp/runtime-memory-iatp-dim-ih/build.gradle.kts
@@ -58,6 +58,9 @@ dependencies {
 
     // for the controller
     implementation(libs.jakarta.rsApi)
+
+    runtimeOnly(libs.edc.lib.store)
+
 }
 
 application {
